requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose
httpx
celery
redis